Rework prebase_regen module to only reset incomplete publish states, clean out db_handler
prebase_regen we previously used to reorder qaqc process runs so that the a particular base run would be picked up as the latest run.
this functionality is no longer needed.

cleaned out db_handler so that only the last few methods which are not fully implemented remain for migration.
cleaned up logger re: removed translate_early_base functionality